backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic>=0.40.0,<1.0.0
python-multipart==0.0.9
beautifulsoup4==4.12.3
aiofiles==23.2.1
python-dotenv==1.0.1
